KL-110/feat: modify ProductUpdateRequestDto
- price 필드 타입 Integer로 수정

+++ src/main/java/taco/klkl/domain/product/dto/request/ProductUpdateRequestDto.java
@@ -7,6 +7,6 @@ public record ProductUpdateRequestDto(
 	Long subcategoryId,
 	Long currencyId,
 	String address,
-	Long price
+	Integer price
 ) {
 }
